top: exploit msg handler override to avoid corrupt tty
Way back in November of 2011, the library was equipped
with an overridable error message handler function. It
was done expressly for a program like top which alters
the tty. But that support was withdrawn shortly after.

This was all done in the lead up to v3.3.2. That's the
release where NLS support was added and it represented
a hectic time. In hindsight, the changes went too far.

So this commit, in a minimal fashion, restores ability
to address a potential fatal library error. After all,
any properly behaving library would never unilaterally
subject a caller to a stderr message and then an exit.

[ when exposing 1 variable in libprocps.sym, 2 other ]
[ existing symbols were repositioned alphabetically. ]

Reference(s):
. generalized library memory provisions
commit 7126cc4491847ce5d50e603fe48666f94bfc60bf
. top exploit library memory provisions
commit 88087ec5a5968304d6461be2c2fbe05885ffdad8
. library xalloc type functions made private
commit 2865ded64e72e0bec721dad9b442dab0899ee2b0
. restored prior top memory logic
commit 05f5deb97c454b3558892ff4266efc81547f88d5

Signed-off-by: Jim Warner <james.warner@comcast.net>