library: eliminated the questionable 'procps.h' header
There was a time when that procps.h file served a more
traditional role. Prior to the commit referenced below
it held just macros plus manifest constants. But, with
that change, such items were replaced with a series of
includes embracing all the library exported functions.

That approach was known to disguise errors which would
have otherwise yielded a compiler warning. And without
such a warning, there was no way to address the error.

So this patch will trade the all inclusive header file
approach for individual includes only where necessary.

Reference(s):
. April 2016, procps.h header file revamped
commit ccb6ae8de14b0cde25b84369ef995bcd69cbf7b6
. Sept 2018, top abandoned use of procps.h
commit a6dfc2382ed1e023dd345cdb1d2388c9b67bcc7f

Signed-off-by: Jim Warner <james.warner@comcast.net>