library: eliminate one potential source of human error
There's a huge toe-stubber awaiting future maintenance
as reflected in that commit below which deals with the
addition of new enumerators to the Item_table. Namely,
whenever the table is grown, one must remember to also
change that existing 'logical_end' enumerator's value.

Well, not anymore! Since that MAXTABLE macro was added
to the procps-private.h header we can now also exploit
it so a 'logical_end' automatically tracks table size.

This change also renders some code associated with the
ITEMTABLE_DEBUG #define unnecessary. So it's gone too.

Reference(s):
. 08/2016, add new enumerators
commit 09e1886c9e731f8b8c89a55d11f72f53f030b2de
. 08/2020, added MAXTABLE macro
commit c865b06c304e645c05eae0f5e06a1f49ea105285
. 08/2020, introduced ITEMTABLE_DEBUG
commit 92d0297e1e4d5946c5b098e37c91c7e524a0eca0

Signed-off-by: Jim Warner <james.warner@comcast.net>